+++ btkey/__init__.py
@@ -3,7 +3,7 @@
 
 import os
 
-__version__ = "0.1.5"
+__version__ = "0.1.6"
 
 
 def from_checkout():